% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LUCscen.R
\name{LUCscen}
\alias{LUCscen}
\title{Generate a new land use change scenario in SWAT+}
\usage{
LUCscen(
  scen.name,
  textInOut.path,
  hru.shape,
  lc,
  lookup.table,
  change.yr,
  skip.copy = FALSE,
  jday = 1,
  fact.diss = 8,
  scen.path = NULL,
  verbose = TRUE,
  exclude.uses = NULL
)
}
\arguments{
\item{scen.name}{Character object containing the name of the new scenario to be created.}

\item{textInOut.path}{Path to the default 'TxtInOut'. The new scenario will be created from this file.}

\item{hru.shape}{Spatial object of type 'SpatVector' that contains the HRUs as obtained from QSWAT+.}

\item{lc}{Raster object of type 'SpatRaster' that contains the land cover for the years to be analysed.}

\item{lookup.table}{A three column data.frame. The first column stores the land use values from the 'lc' object. The
second column stores the SWAT_CODE as in the lookup table that is given to SWAT+. Finally, the third column represents
the land use (lum) identifiers from SWAT+ that can be obtained from the 'landuse.lum' file.}

\item{change.yr}{Character object containing the yera(s) when the land use changes take place (in chronological order).}

\item{jday}{Numerical value that indicates the julian day that the land use changes will take place (default jday = 1).}

\item{scen.path}{Path to the directory where the new scenatio will be created. If not given, the new scenario will be
created in the folder of the default scenario.}

\item{verbose}{Should the progress messages be printed?}
}
\value{

}
\description{
Generate a new land use change scenario in SWAT+
}
