% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_changes.R
\name{plot_changes}
\alias{plot_changes}
\title{Plot the changes of the given land use layers as pie charts}
\usage{
plot_changes(df.changes, ncolumns, df.codes)
}
\arguments{
\item{df.changes}{Object of type 'data.frame' obtained from
applying the function 'extract_changes'.}

\item{ncolumns}{Number of columns of the resulting figure.}

\item{df.codes}{Data frame of two columns. The first column 'ID'
must contain the codes of the land uses, while the second column
'names' contain the respective land cover names.}
}
\value{
This function returns a 3D pie plot for every land cover layer
   given to the 'extract_changes' function.
}
\description{
Plot the changes of the given land use layers as pie charts
}
\author{
Oscar M. Baez-Villanueva
}
