% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/products2stations.R
\name{prec2stations}
\alias{prec2stations}
\title{Convert every grid-cell of a gridded pracipitation product into a station for SWAT+}
\usage{
prec2stations(product, dem, output, hru.shape, start)
}
\arguments{
\item{product}{Precipitation product as 'SpatRast' that includes all the days of the study period as layers.}

\item{dem}{Digital elevation model to be used to extract the elevation of the stations.}

\item{output}{Folder to write the station files (data and metadata).}

\item{hru.shape}{Spatial object of type 'SpatVector' that contains the HRUs as obtained from QSWAT+}

\item{start}{Character object that contains the starting date of the model in the format '%Y%m%d' e.g., '19810101'.}
}
\value{

}
\description{
Convert every grid-cell of a gridded pracipitation product into a station for SWAT+
}
